add modules, fix parcel
