+++ package.json
@@ -7,9 +7,10 @@
   "description": "Yandex Praktikum sprint1 (chat)",
   "main": "index.js",
   "scripts": {
-    "clean": "rm dist/bundle.js",
-    "start": "parcel static/profile.html",
-    "build-prod": "parcel build static/index.html"
+    "start": "npm run clear && parcel build ./static/*.html --no-source-maps --no-cache --public-url ./ && node ./app.js",
+    "dev": "npm run clear && parcel ./static/*.html",
+    "build": "npm run clear && parcel build ./static/*.html --no-source-maps --no-cache --public-url ./",
+    "clear": "rm -rf dist .cache"
   },
   "repository": {
     "type": "git",
@@ -28,6 +29,7 @@
     "normalize.css": "^8.0.1",
     "nunjucks": "^3.2.3",
     "parcel-bundler": "^1.12.5",
+    "parcel-plugin-nunjucks": "^2.2.2",
     "path": "^0.12.7"
   },
   "devDependencies": {
