+++ static/chat-page/chat-cell.html
@@ -1,0 +1,13 @@
+<div class="cell">
+    <div class="badge__avatar">
+        <!-- <img src="" alt="avatar" width="47px" height="47px"> -->
+    </div>
+    <div class="cell__text">
+        <span class="badge__name">Андрей</span>
+        <span class="cell__paragraph">И Human Interface Guidelines и Material Design рекомендуют...</span>
+    </div>
+    <div class="cell__info">
+        <span class="cell__time">14:88</span>
+        <div class="cell__msg-count">4</div>
+    </div>
+</div>
