+++ static/chat-page/chat-msg.html
@@ -1,0 +1,32 @@
+
+      <div class="chat__wrap chat__wrap_get">
+          <div class="chat__text">
+              <p>
+                  Привет! Смотри, тут всплыл интересный кусок лунной космической истории — НАСА в какой-то
+                  момент
+                  попросила Хассельблад адаптировать модель SWC для полетов на Луну. Сейчас мы все знаем что
+                  астронавты летали с моделью 500 EL — и к слову говоря, все тушки этих камер все еще
+                  находятся на
+                  поверхности Луны, так как астронавты с собой забрали только кассеты с пленкой.
+              </p>
+              <p>
+                  Хассельблад в итоге адаптировал SWC для космоса, но что-то пошло не так и на ракету они так
+                  никогда и не попали. Всего их было произведено 25 штук, одну из них недавно продали на
+                  аукционе
+                  за 45000 евро.
+              </p>
+          </div>
+      </div>
+
+    <div class="chat__wrap chat__wrap_send">
+        <p class="chat__text">
+            Хассельблад в итоге адаптировал SWC для космоса, но что-то пошло не так и на ракету они так
+            никогда и не попали. Всего их было произведено 25 штук, одну из них недавно продали на аукционе
+            за 45000 евро.
+        </p>
+    </div>
+
+<div class="chat__wrap chat__wrap_date">
+    19 мортабля
+</div>
+
