+++ static/chat-page/chat-page.less
@@ -1,0 +1,168 @@
+@import "../style/variables";
+.toolbar {
+  padding: 14px 0;
+  display: flex;
+  flex-flow: column;
+  border-bottom: 1px solid @lightgray;
+
+  &__btn {
+    color: @gray;
+    font-size: 12px;
+    font-weight: 500;
+    align-self: flex-end;
+    padding: 14px 0;
+    cursor: pointer;
+  }
+
+  &__search {
+    border-radius: 5px;
+  }
+}
+
+.cell {
+  display: flex;
+  justify-content: space-between;
+  padding: 12px 0;
+  border-bottom: 1px solid @lightgray;
+  align-items: center;
+
+  &__text {
+    display: flex;
+    flex-flow: column;
+    margin-left: 10px;
+    margin-right: 15px;
+    max-width: 193px;
+  }
+
+  &__paragraph {
+    color: @gray;
+    font-size: 12px;
+  }
+
+  &__time {
+    font-size: 10px;
+    color: @gray;
+  }
+
+
+
+  &__msg-count {
+    height: 20px;
+    width: 20px;
+    font-size: 11px;
+    background-color: @brightblue;
+    border-radius: 50px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 5px;
+  }
+}
+
+.content {
+  display: flex;
+  justify-content: space-between;
+  flex-flow: column;
+  align-items: flex-start;
+  padding: 10px 20px;
+
+  &__info-msg {
+    color: @gray;
+    font-weight: 500;
+    margin: auto;
+    text-align: center;
+  }
+
+  &__info-panel {
+    border-bottom: 1px solid @lightblue;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    padding-bottom: 10px;
+  }
+
+  &__chat {
+    display: flex;
+    flex-flow: column;
+    width: 100%;
+    height: 100%;
+    padding-top: 12px;
+  }
+
+  &__msg-panel {
+    border-top: 1px solid @lightblue;
+    width: 100%;
+    padding: 15px 0 5px 0;
+    display: flex;
+    justify-content: space-between;
+  }
+
+}
+
+.info-panel {
+  &__card {
+    display: flex;
+    flex-flow: row;
+    align-items: center;
+
+    p {
+      margin-left: 10px;
+    }
+  }
+
+  &__avatar {
+    width: 34px;
+    height: 34px;
+    border-radius: 50px;
+    background-color: @darkblue;
+  }
+
+  &__name {
+    margin-left: 10px;
+  }
+
+  &__menu {
+    display: flex;
+    flex-flow: column;
+    align-items: flex-start;
+    min-width: 200px;
+    padding: 12px;
+    position: absolute;
+    top: 55px;
+    right: 0;
+    z-index: 5;
+    background-color: @darkblue;
+    border-radius: 12px;
+    border: 1px solid @lightgray;
+
+    &_icon {
+      border: 2px solid @lightblue;
+      border-radius: 50px;
+      padding: 3px;
+      margin-right: 6px;
+    }
+  }
+
+  &__menu-btn {
+    width: 25px;
+    height: 25px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    position: relative;
+  }
+}
+
+.msg-panel {
+  &__search {
+    width: 88%;
+    border-radius: 30px;
+    margin: 0 10px;
+  }
+
+  &__add-file {
+    position: relative;
+  }
+}
