--- static/chat.html
@@ -1,1 +1,0 @@
-<p class="">chat</p>
