+++ static/index.js
@@ -1,8 +1,1 @@
-// import './chat.js'
-const nunjucks = require('nunjucks')
-
-
-nunjucks.configure({ autoescape: true });
-nunjucks.renderString('Hello {{ username }}', { username: 'James' });
-
-
+console.log('###### >> index.js loaded:');
