+++ static/login-page.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Colloquy Chat</title>
-    <link rel="stylesheet" href="./style/styles.less">
+    <link rel="stylesheet" href="./style/styles.css">
 </head>
 <body>
     <div class="root">
