+++ static/style/styles.css
@@ -1,27 +1,3 @@
-body {
-  background-color: #1B2E42;
-  color: #FFFAFA;
-  font-family: 'Montserrat', sans-serif;
-  font-weight: 500;
-}
-.root {
-  max-width: 1280px;
-  min-height: 80vh;
-  display: flex;
-  justify-content: center;
-}
-.main {
-  display: grid;
-  grid-template-columns: 1fr 3fr;
-  align-self: auto;
-  border: 1px solid rgba(138, 196, 255, 0.3);
-  border-radius: 5px;
-}
-.sidebar {
-  border-right: 1px solid rgba(138, 196, 255, 0.3);
-  width: 320px;
-  padding: 10px;
-}
 .toolbar {
   padding: 14px 0;
   display: flex;
@@ -44,6 +20,7 @@ body {
   justify-content: space-between;
   padding: 12px 0;
   border-bottom: 1px solid rgba(138, 196, 255, 0.3);
+  align-items: center;
 }
 .cell__text {
   display: flex;
@@ -92,45 +69,210 @@ body {
   width: 100%;
   padding-bottom: 10px;
 }
-.content__chatfield {
+.content__chat {
   display: flex;
   flex-flow: column;
   width: 100%;
   height: 100%;
+  padding-top: 12px;
 }
 .content__msg-panel {
   border-top: 1px solid #8AC4FF;
   width: 100%;
-  padding: 10px 0;
+  padding: 15px 0 5px 0;
   display: flex;
   justify-content: space-between;
 }
 .info-panel__card {
   display: flex;
   flex-flow: row;
+  align-items: center;
 }
 .info-panel__card p {
   margin-left: 10px;
 }
+.info-panel__avatar {
+  width: 34px;
+  height: 34px;
+  border-radius: 50px;
+  background-color: #364f6b;
+}
 .info-panel__name {
   margin-left: 10px;
 }
-.info-panel__btn {
+.info-panel__menu {
+  display: flex;
+  flex-flow: column;
+  align-items: flex-start;
+  min-width: 200px;
+  padding: 12px;
+  position: absolute;
+  top: 55px;
+  right: 0;
+  z-index: 5;
+  background-color: #364f6b;
+  border-radius: 12px;
+  border: 1px solid rgba(138, 196, 255, 0.3);
+}
+.info-panel__menu_icon {
+  border: 2px solid #8AC4FF;
+  border-radius: 50px;
+  padding: 3px;
+  margin-right: 6px;
+}
+.info-panel__menu-btn {
   width: 25px;
   height: 25px;
   display: flex;
   justify-content: center;
   align-items: center;
   cursor: pointer;
+  position: relative;
 }
 .msg-panel__search {
-  width: 80%;
+  width: 88%;
   border-radius: 30px;
   margin: 0 10px;
 }
+.msg-panel__add-file {
+  position: relative;
+}
+.modal-options {
+  display: flex;
+  flex-flow: column;
+  align-items: flex-start;
+  min-width: 160px;
+  position: absolute;
+  bottom: 65px;
+  padding: 15px;
+  border-radius: 12px;
+  border: 1px solid rgba(138, 196, 255, 0.3);
+  color: #FFFAFA;
+  background-color: #364f6b;
+  box-shadow: 0 0 6px rgba(0, 0, 0, 0.14);
+  z-index: 5;
+}
+.modal-options__point {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  padding: 8px;
+  cursor: pointer;
+  font-size: 12px;
+}
+.modal-options svg {
+  margin-right: 8px;
+}
+.modal {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  z-index: 6;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.4);
+  /* Black w/ opacity */
+}
+.modal__content {
+  background-color: #364f6b;
+  border: 1px solid rgba(138, 196, 255, 0.3);
+  display: flex;
+  flex-flow: column;
+  justify-content: center;
+  align-items: center;
+  min-width: 320px;
+  padding: 40px;
+  border-radius: 8px;
+}
+.modal__link {
+  color: #8AC4FF;
+  font-size: 12px;
+  text-decoration: #8AC4FF underline;
+  cursor: pointer;
+  margin: 20px 0;
+}
+body {
+  background-color: #1B2E42;
+  color: #FFFAFA;
+  font-family: 'Helvetica', sans-serif;
+  font-weight: 500;
+}
+input {
+  background-color: #1B2E42;
+  border: none;
+  outline: none;
+  color: #FFFAFA;
+  padding: 5px;
+}
+input-webkit-autofill {
+  background-color: #999999;
+}
+input:-webkit-autofill,
+input:-webkit-autofill:hover,
+input:-webkit-autofill:focus,
+input:-webkit-autofill:active {
+  transition: background-color 5000s ease-in-out 0s;
+  -webkit-text-fill-color: #FFFAFA;
+}
+button {
+  border: none;
+  outline: none;
+  cursor: pointer;
+}
+.root {
+  max-width: 1280px;
+  min-height: 97vh;
+  display: flex;
+  justify-content: center;
+  margin: auto;
+}
+.main {
+  display: grid;
+  grid-template-columns: 1fr 3fr;
+  align-self: auto;
+  border: 1px solid rgba(138, 196, 255, 0.3);
+  border-radius: 5px;
+}
+.sidebar {
+  border-right: 1px solid rgba(138, 196, 255, 0.3);
+  width: 320px;
+  padding: 10px;
+}
+.chat__wrap {
+  padding: 12px;
+  border-radius: 12px;
+  width: 50%;
+  margin-bottom: 12px;
+}
+.chat__wrap_get {
+  background-color: #2B5278;
+  border-top-left-radius: 0;
+}
+.chat__wrap_send {
+  background-color: #1067BC;
+  align-self: flex-end;
+  border-bottom-right-radius: 0;
+}
+.chat__wrap_date {
+  color: #999999;
+  align-self: center;
+  font-size: 12px;
+  text-align: center;
+}
+.chat__text {
+  font-size: 12px;
+  line-height: 15px;
+  margin: 0;
+}
+.chat__text p {
+  margin: 0;
+}
 .input__search {
   border: none;
-  background-color: #364F6B;
+  background-color: #364f6b;
   color: #FFFAFA;
   outline: none;
   padding: 10px;
@@ -139,13 +281,129 @@ body {
   color: #999999;
   font-weight: 500;
 }
+.input__search:focus {
+  box-shadow: 0 0 5px #0064b3;
+}
+.button {
+  background-color: transparent;
+  border: 0;
+  outline: 0;
+  cursor: pointer;
+}
 .badge__name {
   font-size: 13px;
-  margin-bottom: 5px;
+  margin: 4px 0;
 }
 .badge__avatar {
   width: 47px;
   height: 47px;
   border-radius: 50px;
-  background-color: #364F6B;
+  background-color: #364f6b;
 }
+.form {
+  display: flex;
+  justify-content: center;
+  flex-flow: column;
+  width: 100%;
+}
+.form-wrap {
+  border-radius: 12px;
+  border: 1px solid rgba(138, 196, 255, 0.3);
+  min-width: 320px;
+  padding: 30px;
+  display: flex;
+  flex-flow: column;
+  justify-content: space-between;
+  align-items: center;
+  margin: auto;
+}
+.form-group {
+  position: relative;
+  width: 100%;
+  margin-bottom: 40px;
+}
+.form-label {
+  font-size: 11px;
+  color: #999999;
+  position: absolute;
+  top: -10px;
+  left: 5px;
+}
+.form-input {
+  border-bottom: 1px solid #8AC4FF;
+  width: 97%;
+}
+.form-input::placeholder {
+  color: #999999;
+}
+.form-btn-submit {
+  padding: 12px;
+  background-color: #8AC4FF;
+  border-radius: 8px;
+}
+.form-btn-option {
+  background-color: transparent;
+  color: #8AC4FF;
+  margin-top: 20px;
+}
+.form-btn-cancel {
+  background-color: #EC4336;
+  border-radius: 8px;
+  padding: 12px;
+}
+.form-error {
+  color: #EC4336;
+  font-size: 11px;
+  position: absolute;
+  top: 32px;
+}
+.sidebar-back {
+  border-right: 1px solid #8AC4FF;
+  display: flex;
+  align-items: center;
+  padding: 18px;
+}
+.sidebar-back__btn {
+  cursor: pointer;
+}
+.profile-content {
+  width: 100%;
+  padding: 60px;
+}
+.profile-avatar {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #364f6b;
+  border-radius: 50%;
+  width: 130px;
+  height: 130px;
+  margin-bottom: 40px;
+}
+.profile-title {
+  margin-bottom: 40px;
+}
+.profile-form-wrap {
+  border: 0;
+}
+.profile-form {
+  align-items: center;
+  max-width: 510px;
+}
+.profile-form-group {
+  margin-bottom: 20px;
+}
+.profile-form-input {
+  text-align: right;
+  color: #999999;
+}
+.profile-form-label {
+  font-size: 13px;
+  top: 6px;
+  color: #FFFAFA;
+}
+.profile-btn {
+  width: 50%;
+  margin-top: 30px;
+}
+/*# sourceMappingURL=styles.css.map */
