+++ static/style/styles.less
@@ -1,5 +1,6 @@
 @import './variables.less';
-
+@import '../chat-page/chat-page.less';
+@import "../modal/modal.less";
 // * {
 //   border: 1px solid pink;
 // }
@@ -57,207 +58,6 @@ button {
   border-right: 1px solid @lightgray;
   width: 320px;
   padding: 10px;
-}
-
-.toolbar {
-  padding: 14px 0;
-  display: flex;
-  flex-flow: column;
-  border-bottom: 1px solid @lightgray;
-
-  &__btn {
-    color: @gray;
-    font-size: 12px;
-    //TODO add font 500
-    font-weight: 500;
-    align-self: flex-end;
-    padding: 14px 0;
-    cursor: pointer;
-  }
-
-  &__search {
-    border-radius: 5px;
-  }
-}
-
-.cell {
-  display: flex;
-  justify-content: space-between;
-  padding: 12px 0;
-  border-bottom: 1px solid @lightgray;
-  align-items: center;
-
-  &__text {
-    display: flex;
-    flex-flow: column;
-    margin-left: 10px;
-    margin-right: 15px;
-    max-width: 193px;
-  }
-
-  &__paragraph {
-    color: @gray;
-    font-size: 12px;
-  }
-
-  &__time {
-    font-size: 10px;
-    color: @gray;
-  }
-
-  &__info {
-
-  }
-
-  &__msg-count {
-    height: 20px;
-    width: 20px;
-    font-size: 11px;
-    background-color: @brightblue;
-    border-radius: 50px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 5px;
-  }
-}
-
-.content {
-  display: flex;
-  justify-content: space-between;
-  flex-flow: column;
-  align-items: flex-start;
-  padding: 10px 20px;
-
-  &__info-msg {
-    color: @gray;
-    font-weight: 500;
-    margin: auto;
-    text-align: center;
-  }
-
-  &__info-panel {
-    border-bottom: 1px solid @lightblue;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    padding-bottom: 10px;
-  }
-
-  &__chat {
-    display: flex;
-    flex-flow: column;
-    width: 100%;
-    height: 100%;
-    padding-top: 12px;
-  }
-
-  &__msg-panel {
-    border-top: 1px solid @lightblue;
-    width: 100%;
-    padding: 15px 0 5px 0;
-    display: flex;
-    justify-content: space-between;
-  }
-
-}
-
-.info-panel {
-  &__card {
-    display: flex;
-    flex-flow: row;
-    align-items: center;
-
-    p {
-      margin-left: 10px;
-    }
-  }
-
-  &__avatar {
-    width: 34px;
-    height: 34px;
-    border-radius: 50px;
-    background-color: @darkblue;
-  }
-
-  &__name {
-    margin-left: 10px;
-  }
-
-  &__menu {
-    display: flex;
-    flex-flow: column;
-    align-items: flex-start;
-    min-width: 200px;
-    padding: 12px;
-    position: absolute;
-    top: 55px;
-    right: 0;
-    z-index: 5;
-    background-color: #1B2E42;
-    border-radius: 12px;
-    border: 1px solid @lightgray;
-
-    &_icon {
-      border: 2px solid @lightblue;
-      border-radius: 50px;
-      padding: 3px;
-      margin-right: 6px;
-    }
-  }
-
-  &__menu-btn {
-    width: 25px;
-    height: 25px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-    position: relative;
-  }
-}
-
-.msg-panel {
-  &__search {
-    width: 88%;
-    border-radius: 30px;
-    margin: 0 10px;
-  }
-
-  &__add-file {
-    position: relative;
-  }
-}
-
-// modal window send file
-.modal-options {
-  display: flex;
-  flex-flow: column;
-  align-items: flex-start;
-  min-width: 160px;
-  position: absolute;
-  bottom: 65px;
-  padding: 15px;
-  border-radius: 12px;
-  border: 1px solid @lightgray;
-  color: @font-color;
-  background-color: @darkblue;
-  box-shadow: 0 0 6px rgba(0, 0, 0, 0.14);
-  z-index: 5;
-
-  &__point {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    padding: 8px;
-    cursor: pointer;
-    font-size: 12px;
-  }
-
-  svg {
-    margin-right: 8px;
-  }
 }
 
 // chat options
@@ -473,37 +273,3 @@ button {
   }
 }
 
-// модальное окно
-.modal {
-  // TODO значально display none
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  z-index: 6;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
-
-  &__content {
-    background-color: @darkblue;
-    border: 1px solid @lightgray;
-    display: flex;
-    flex-flow: column;
-    justify-content: center;
-    align-items: center;
-    min-width: 320px;
-    padding: 40px;
-    border-radius: 8px;
-  }
-  &__link {
-    color: @lightblue;
-    font-size: 12px;
-    text-decoration: @lightblue underline;
-    cursor: pointer;
-    margin: 20px 0;
-  }
-
-}
