+++ static/testData.js
@@ -1,0 +1,6 @@
+const data = {
+  name: 'Геворг',
+  foo: 'fffffffooooloaded'
+}
+
+module.exports = data;
